Fixes issues with worker startup
