--- Freshli.Web.Worker/Worker.cs
@@ -1,26 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-
-namespace Freshli.Web.Worker {
-  public class Worker : BackgroundService {
-    private readonly ILogger<Worker> _logger;
-
-    public Worker(ILogger<Worker> logger) {
-      _logger = logger;
-    }
-
-    protected override async Task ExecuteAsync(
-      CancellationToken stoppingToken
-    ) {
-      while (!stoppingToken.IsCancellationRequested) {
-        _logger.LogInformation("Worker running at: {time}", DateTimeOffset.Now);
-        await Task.Delay(1000, stoppingToken);
-      }
-    }
-  }
-}
